Non-default team detection & reset button disable

+++ template-engine/private/src/routing/reset.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { deleteUser, createUser } = require('../user');
 const { validateFile } = require('../utils');
 const { getUserTeamsListJSONPath } = require('../userPath');
-const { deleteTeam, cloneTeamFromDefault, copyTeamListTeam } = require('../teamStorage');
+const { deleteTeam, cloneTeamFromDefault, copyTeamListTeam, hasTeamDefault } = require('../teamStorage');
 
 const router = express.Router();
 
@@ -25,6 +25,9 @@ router.put('/:teamId', (req, res) => {
   const { teamId } = req.params;
   console.log(`Resetting team ${teamId} from ${username}`);
   try {
+    if (!hasTeamDefault(username, teamId)) {
+      throw new Error('Team is not resettable');
+    }
     deleteTeam(username, teamId);
     const defaultUsername = 'default';
     cloneTeamFromDefault(username, teamId);
